fix: update tests
