+++ src/components/__tests__/BottomNavigation.test.tsx
@@ -29,58 +29,6 @@ declare global {
   }
 }
 
-type AnimatedTiming = (
-  value: Animated.Value | Animated.ValueXY,
-  config: Animated.TimingAnimationConfig
-) => Animated.CompositeAnimation;
-
-type AnimatedParallel = (animations: Array<Animated.CompositeAnimation>) => {
-  start: (callback?: Animated.EndCallback) => void;
-};
-
-// Make sure any animation finishes before checking the snapshot results
-jest.mock('react-native', () => {
-  const RN = jest.requireActual('react-native');
-
-  const timing: AnimatedTiming = (value, config) => ({
-    start: (callback) => {
-      value.setValue(config.toValue as any);
-      callback?.({ finished: true });
-    },
-    value,
-    config,
-    stop: () => {
-      throw new Error('Not implemented');
-    },
-    reset: () => {
-      throw new Error('Not implemented');
-    },
-  });
-  RN.Animated.timing = timing;
-
-  const parallel: AnimatedParallel = (animations) => ({
-    start: (callback) => {
-      const results = animations.map((animation) => {
-        animation.start();
-        return { finished: true };
-      });
-      callback?.({ finished: results.every((result) => result.finished) });
-    },
-  });
-
-  RN.Animated.parallel = parallel;
-
-  RN.Dimensions.get = () => ({
-    fontScale: 1,
-  });
-
-  return RN;
-});
-
-jest.mock('react-native-safe-area-context', () => ({
-  useSafeAreaInsets: () => ({ bottom: 0, left: 0, right: 0, top: 0 }),
-}));
-
 const icons = ['magnify', 'camera', 'inbox', 'heart', 'shopping-music'];
 
 const createState = (index: number, length: number) => ({
