+++ src/components/__tests__/DataTable.test.tsx
@@ -1,12 +1,9 @@
 import * as React from 'react';
 
 import { render } from '@testing-library/react-native';
-import mockSafeAreaContext from 'react-native-safe-area-context/jest/mock';
 
 import Checkbox from '../Checkbox';
 import DataTable from '../DataTable/DataTable';
-
-jest.mock('react-native-safe-area-context', () => mockSafeAreaContext);
 
 describe('DataTable.Header', () => {
   it('renders data table header', () => {
