+++ src/components/__tests__/Dialog.test.tsx
@@ -12,10 +12,6 @@ import { act, fireEvent, render } from '@testing-library/react-native';
 import Dialog from '../../components/Dialog/Dialog';
 import Button from '../Button/Button';
 
-jest.mock('react-native-safe-area-context', () => ({
-  useSafeAreaInsets: () => ({ bottom: 44, left: 0, right: 0, top: 37 }),
-}));
-
 jest.mock('react-native/Libraries/Utilities/BackHandler', () =>
   // eslint-disable-next-line jest/no-mocks-import
   require('react-native/Libraries/Utilities/__mocks__/BackHandler')
