+++ src/components/__tests__/FABGroup.test.tsx
@@ -8,10 +8,6 @@ import { getTheme } from '../../core/theming';
 import FAB from '../FAB';
 import { getFABGroupColors } from '../FAB/utils';
 
-jest.mock('react-native-safe-area-context', () => ({
-  useSafeAreaInsets: () => ({ bottom: 34, left: 0, right: 0, top: 47 }),
-}));
-
 describe('getFABGroupColors - backdrop color', () => {
   it('should return custom color', () => {
     expect(
